+++ tests/test_multipoint.py
@@ -1,5 +1,7 @@
 """Test MultiPoint."""
 
+import math
+
 import pytest
 
 from pygeoif import geometry
@@ -156,12 +158,18 @@ def test_empty() -> None:
 
 
 def test_repr_empty() -> None:
-    multipoint = geometry.MultiPoint([(None, None)])
+    multipoint = geometry.MultiPoint([(math.nan, math.nan)])
 
-    assert repr(multipoint) == "MultiPoint(((),))"
+    assert repr(multipoint) == "MultiPoint(())"
 
 
 def test_empty_bounds() -> None:
     multipoint = geometry.MultiPoint([(None, None)])
 
     assert multipoint.bounds == ()
+
+
+def test_empty_geoms() -> None:
+    multipoint = geometry.MultiPoint([(math.nan, math.nan)])
+
+    assert not list(multipoint.geoms)